LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/smooth-scroll-nonstop-expected.txt


FAIL noop scrollTo doesn't interrupt ongoing smooth scroll. assert_equals: scroller reached the target offset expected 400 but got 53
FAIL noop scrollIntoView doesn't interrupt ongoing smooth scroll. assert_equals: scroller reached the target offset expected 400 but got 51
